refactor: self-update
