+++ tools/plugin-commands-self-updater/src/index.ts
@@ -1,4 +1,4 @@
 import * as selfUpdate from './selfUpdate'
+export { installPnpmToTools } from './installPnpmToTools'
 
 export { selfUpdate }
-export { installPnpmToTools } from './selfUpdate'
